+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1902.4.0]
+
+### Changed
+* **IMPORTANT**: cross-mod integration has been removed from FTB Chunks in this release!
+  * Install the **FTB XMod Compat** mod to reintroduce integration with the KubeJS, Game Stages (forge), Waystones and Common Protection API (fabric) mods
+  * The rationale behind this change is to move dependencies on external (non-FTB) mods into one place to minimise future disruption due to API-breaking changes (e.g. KubeJS 6.0 -> 6.1)
+
 ## [1902.3.23]
 
 ### Added